Clone repository instead of checking out the sources
"git checkout" changed the index of the original repository, which was
undesired (to put it mildly).

Shared clone adds 4.5 megabytes per repository (mostly for the index),
but it opens new possibilities, such as cleaning with "git clean",
support for "git annotate" for any file on any tree, the ability to
patch the source tree and revert to the pristine sources and the ability
to update the source trees (especially important for linux-next).