hardlink -t doesn't work on Fedora 21, use hardlink -c
hardlink -c ignores mode and ownership. Ignoring mode is potentially
risky, as it could unset the executable flag on some script that is
non-executalbe on other kernels. But the build system doesn't seem to
rely on the executable bit in the source tree with a few exceptions that
don't matter for module compilation.